+++ frontends/main/src/app-pages/PodcastPage/PodcastsListingPage/constants.ts
@@ -1,0 +1,3 @@
+export const EPISODES_PAGE_SIZE = 5
+export const SERIES_FEATURED_COUNT = 2
+export const SERIES_MORE_COUNT = 5
