+++ frontends/main/src/app-pages/PodcastPage/PodcastsListingPage/index.ts
@@ -1,0 +1,1 @@
+export { PodcastsListingPage, default } from "./PodcastsListingPage"
